Created Profile pages for users (see Description for more detail.)
The cards that are seen for results now has the username in place of the user first name. The username are not clickable and it allows users to navigate to the posters user profile page. The user profile page will display all the courses the user tutors for if there are any.