Message page now populates. See Description.
Logged in users can now select the message link in the nav bar to access all messages sent to them. The page and partials hbs was created to populate on the frontend. The header.hbs and index.js has been updated to handle navigation. Messages has a model and middleware created to handle messages.
The user.js has ben been updated to remove extra stuff.