Dealt with merge issues (see description for more context)
login.js using added an alert to allow users to know if there was an issue logging in. The file was also formatted.
Registration.js had unused code commented out that was causing things to not work. The file was also reformatted.
color_book.css had an attribute in buttons to change because it was causing weird display.
style.css was removed since it is not being used.
home.hbs had bootstrap added to it.
login.hbs was reformatted
register.hbs had a button changed to input due to an issue with submission.
route/user.js had a flash commented out since in current state, flash is not enabled.