Separated search results to its own independent page (see description)
frontend.js underwent an overhaul due to the cards no longer being generated there for populating search results. A lot of the code was removed such as the CreateCard and hover over the return to home text functions. The ExecuteSearch function was simplified to a reassign the window location based on search parameters.
cards.hbs needed to be adjusted to remain persistent in showing thumbnails.
hearders.hbs was overhauled to remove the navbar appear conditional and needed to allow for pre-existing query to populate in search bar.
search_results.hbs was modified to satisfy needs, but mirrors the original landing page.
app.js needed to have the GetAllPostCoursesPrefix be functional on all pages for the navbar.
route/index.js needed the function removed from the home route for it was relocated.
route/post.js had the /search modified to render the search result page.